# public/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse": {
				"properties": {
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT Authorization header using the Bearer scheme."
			}
		}
	},
	"info": {
		"title": "boiler_api_5300",
		"version": "1.0.0",
		"description": "Boilerplate apis",
		"license": {
			"name": "license"
		},
		"contact": {
			"name": "STS"
		}
	},
	"paths": {
...
		"/reservation/create_reservation": {
			"post": {
				"operationId": "CreateReservation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"reservation Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"reserved_date": {
										"type": "string",
										"format": "date-time"
									},
									"totalBook_reserved": {
										"type": "number",
										"format": "double"
									},
									"bookId": {
										"type": "string"
									},
									"member_id": {
										"type": "string"
									}
								},
								"required": [
									"reserved_date",
									"totalBook_reserved",
									"bookId",
									"member_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/reservation/get_reservation_book": {
			"get": {
				"operationId": "GetreservedBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"reservation Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/reservation/update_reservation_book": {
			"put": {
				"operationId": "UpdatereservedBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"reservation Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"totalBook_reserved": {
										"type": "number",
										"format": "double"
									},
									"book_name": {
										"type": "string"
									},
									"_id": {
										"type": "string"
									}
								},
								"required": [
									"totalBook_reserved",
									"book_name",
									"_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/reservation/delete_reservation_book": {
			"delete": {
				"operationId": "DeletereservedBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"reservation Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"_id": {
										"type": "string"
									}
								},
								"required": [
									"_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/user/borrow/add_borrow": {
			"post": {
				"operationId": "AddBorrow",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "API: Add Borrow (Issue Book)\r\nMethod: POST\r\nURL: /add_borrow\r\n\r\nBody:\r\n{\r\n  \"member_id\": \"ObjectId\",\r\n  \"book_id\": \"ObjectId\",\r\n  \"issueDate\": 20260411,\r\n  \"lastDate\": 20260420\r\n}",
				"tags": [
					"Borrow Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"lastDate": {
										"type": "string"
									},
									"issueDate": {
										"type": "string"
									},
									"book_id": {
										"type": "string"
									},
									"member_id": {
										"type": "string"
									}
								},
								"required": [
									"lastDate",
									"issueDate",
									"book_id",
									"member_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/borrow/get_borrow_book": {
			"get": {
				"operationId": "GetBorrowBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Borrow Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/borrow/update_borrow_book": {
			"put": {
				"operationId": "Update_borrow_book",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Borrow Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"issueDate": {
										"type": "string"
									},
									"renewCount": {
										"type": "number",
										"format": "double"
									},
									"status": {
										"type": "boolean"
									},
									"lastDate": {
										"type": "string"
									},
									"_id": {
										"type": "string"
									}
								},
								"required": [
									"issueDate",
									"renewCount",
									"status",
									"lastDate",
									"_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/borrow/delete_borrow_book": {
			"delete": {
				"operationId": "Delete_borrow_book",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Borrow Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"_id": {
										"type": "string"
									}
								},
								"required": [
									"_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}